requirements.txt
toml~=0.10.2
numpy~=2.1.0
scikit-learn~=1.5.2
flask~=3.0.3
pillow~=10.4.0
streamlit~=1.38.0